팝업창 설정 완료, 게시글 스크롤이 안되는 버그 수정 완료.
